Restructuring, docs
