+++ language-plutus-core/src/Language/PlutusCore/Constant/Dynamic/OffChain.hs
@@ -13,6 +13,7 @@ import           Language.PlutusCore.Constant.Function
 import           Language.PlutusCore.Constant.Typed
 import           Language.PlutusCore.Core
 import           Language.PlutusCore.Error
+import           Language.PlutusCore.MkPlc
 import           Language.PlutusCore.Name
 import           Language.PlutusCore.Quote
 import           Language.PlutusCore.TypeCheck
