+++ language-plutus-core/src/Language/PlutusCore/Core/Type.hs
@@ -7,8 +7,7 @@
 {-# LANGUAGE UndecidableInstances  #-}
 
 module Language.PlutusCore.Core.Type
-    ( Gas(..)
-    , Kind(..)
+    ( Kind(..)
     , Type(..)
     , BuiltinName(..)
     , DynamicBuiltinName(..)
@@ -18,6 +17,7 @@ module Language.PlutusCore.Core.Type
     , Value
     , Version(..)
     , Program(..)
+    , UniOf
     , Normalized(..)
     , HasUniques
     , defaultVersion
@@ -44,10 +44,6 @@ import           Instances.TH.Lift            ()
 Equality of two things does not depend on their annotations.
 So don't use @deriving Eq@ for things with annotations.
 -}
-
-newtype Gas = Gas
-    { unGas :: Natural
-    }
 
 data Kind ann
     = Type ann
@@ -134,6 +130,11 @@ data Version ann
 -- | A 'Program' is simply a 'Term' coupled with a 'Version' of the core language.
 data Program tyname name uni ann = Program ann (Version ann) (Term tyname name uni ann)
     deriving (Show, Functor, Generic, NFData, Lift, Hashable)
+
+-- | Extract the universe from a type.
+type family UniOf a :: * -> *
+
+type instance UniOf (Term tyname name uni ann) = uni
 
 newtype Normalized a = Normalized
     { unNormalized :: a
